BJS2-40683-Приглашение присоединиться к этапу. Исправления: подправил основную и тестовую структуру проекта для StageInvitation.
