+++ src/main/java/faang/school/projectservice/filters/stageinvitation/InvitationDescriptionFilterImpl.java
@@ -1,7 +1,6 @@
-package faang.school.projectservice.filters.impl;
+package faang.school.projectservice.filters.stageinvitation;
 
 import faang.school.projectservice.dto.invitation.StageInvitationFilterDto;
-import faang.school.projectservice.filters.abstracts.StageInvitationFilter;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
 import faang.school.projectservice.model.stage_invitation.StageInvitationStatus;
 import org.springframework.stereotype.Component;
