+++ src/main/java/faang/school/projectservice/filters/stageinvitation/InvitationStatusFilterImpl.java
@@ -1,23 +1,22 @@
-package faang.school.projectservice.filters.impl;
+package faang.school.projectservice.filters.stageinvitation;
 
 import faang.school.projectservice.dto.invitation.StageInvitationFilterDto;
-import faang.school.projectservice.filters.abstracts.StageInvitationFilter;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
 import org.springframework.stereotype.Component;
 
 import java.util.stream.Stream;
 
 @Component
-public class InvitationAuthorFilterImpl implements StageInvitationFilter {
+public class InvitationStatusFilterImpl implements StageInvitationFilter {
     @Override
     public boolean isApplicable(StageInvitationFilterDto filterDto) {
-        return filterDto.getAuthorId() != null;
+        return filterDto.getStatus() != null;
     }
 
     @Override
     public Stream<StageInvitation> apply(Stream<StageInvitation> invitationStream,
                                          StageInvitationFilterDto filterDto) {
         return invitationStream.filter(stageInvitation ->
-                stageInvitation.getAuthor().getId().equals(filterDto.getAuthorId()));
+                stageInvitation.getStatus().equals(filterDto.getStatus()));
     }
 }
