+++ src/main/java/faang/school/projectservice/filters/stageinvitation/StageInvitationFilter.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.filters.abstracts;
+package faang.school.projectservice.filters.stageinvitation;
 
 import faang.school.projectservice.dto.invitation.StageInvitationFilterDto;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
