+++ src/main/java/faang/school/projectservice/validator/stageinvitation/StageInvitationValidatorImpl.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.validator.impl;
+package faang.school.projectservice.validator.stageinvitation;
 
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.jpa.StageInvitationJpaRepository;
@@ -6,7 +6,6 @@ import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage.Stage;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
 import faang.school.projectservice.model.stage_invitation.StageInvitationStatus;
-import faang.school.projectservice.validator.abstracts.StageInvitationValidator;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
