+++ src/test/java/faang/school/projectservice/filters/stageinvitation/InvitationAuthorFilterImplTest.java
@@ -1,8 +1,8 @@
-package faang.school.projectservice.filters.impl;
+package faang.school.projectservice.filters.stageinvitation;
 
 import faang.school.projectservice.dto.invitation.StageInvitationFilterDto;
+import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
-import faang.school.projectservice.model.stage_invitation.StageInvitationStatus;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -15,18 +15,18 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(MockitoExtension.class)
-class InvitationDescriptionFilterImplTest {
-    private static final String TEST_DESCRIPTION = "test";
+class InvitationAuthorFilterImplTest {
+    private static final Long AUTHOR_ID = 1L;
 
-    private final InvitationDescriptionFilterImpl descriptionFilter = new InvitationDescriptionFilterImpl();
+    private final InvitationAuthorFilterImpl authorFilter = new InvitationAuthorFilterImpl();
 
     @Test
     public void testIsApplicable() {
         StageInvitationFilterDto filterDto = StageInvitationFilterDto.builder()
-                .description(TEST_DESCRIPTION)
+                .authorId(1L)
                 .build();
 
-        boolean isApplicable = descriptionFilter.isApplicable(filterDto);
+        boolean isApplicable = authorFilter.isApplicable(filterDto);
 
         assertTrue(isApplicable);
     }
@@ -35,29 +35,33 @@ class InvitationDescriptionFilterImplTest {
     public void testIsNotApplicable() {
         StageInvitationFilterDto filterDto = new StageInvitationFilterDto();
 
-        boolean isApplicable = descriptionFilter.isApplicable(filterDto);
+        boolean isApplicable = authorFilter.isApplicable(filterDto);
 
         assertFalse(isApplicable);
     }
 
     @Test
-    public void testApplyDescriptionFilter() {
+    public void testApplyAuthorFilter() {
         StageInvitationFilterDto filterDto = StageInvitationFilterDto.builder()
-                .description(TEST_DESCRIPTION)
+                .authorId(AUTHOR_ID)
+                .build();
+        TeamMember firstMember = TeamMember.builder()
+                .id(AUTHOR_ID)
+                .build();
+        TeamMember secondMember = TeamMember.builder()
+                .id(AUTHOR_ID + 1)
                 .build();
         StageInvitation firstInvitation = StageInvitation.builder()
-                .status(StageInvitationStatus.REJECTED)
-                .description("test123")
+                .author(firstMember)
                 .build();
         StageInvitation secondInvitation = StageInvitation.builder()
-                .status(StageInvitationStatus.REJECTED)
-                .description("123")
+                .author(secondMember)
                 .build();
         Stream<StageInvitation> invitationStream = Stream.of(firstInvitation, secondInvitation);
 
-        List<StageInvitation> resultList = descriptionFilter.apply(invitationStream, filterDto).toList();
+        List<StageInvitation> resultList = authorFilter.apply(invitationStream, filterDto).toList();
 
         assertEquals(1, resultList.size());
-        assertTrue(resultList.get(0).getDescription().contains(TEST_DESCRIPTION));
+        assertEquals(AUTHOR_ID, resultList.get(0).getAuthor().getId());
     }
 }
