+++ src/test/java/faang/school/projectservice/filters/stageinvitation/InvitationStatusFilterImplTest.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.filters.impl;
+package faang.school.projectservice.filters.stageinvitation;
 
 import faang.school.projectservice.dto.invitation.StageInvitationFilterDto;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
